A new styles.
A new "separator" class - paragraph class that replaces the html element "horizontal rule" - <hr>. After its application, the paragraph becomes a horizontal line of 1 pixel height in white. This class can be used only for paragraphs placed in content blocks - (see example 3). You can still use for paragraph the divider class derived from Materialize CSS, but such an element is hard to click to pint or select in the Summernote (Lite) editor, because it is only 1 pixel high (!). A new classes to style properly the content in the Summernote Lite content editing area.